Make per-entity subscriptions available to agents
