+++ backend/api/entity.go
@@ -102,6 +102,8 @@ func (e *EntityClient) UpdateEntity(ctx context.Context, entity *corev2.Entity) 
 		}
 	} else {
 		config, _ := corev3.V2EntityToV3(entity)
+		// Ensure per-entity subscription does not get removed
+		config.Subscriptions = corev2.AddEntitySubscription(config.Metadata.Name, config.Subscriptions)
 		req := storev2.NewResourceRequestFromResource(ctx, config)
 
 		wConfig, err := wrap.Resource(config)
